rename various for semantic purposes and add comments
